requirements.txt
lxml==4.9.0
pdf2image==1.16.0
Pillow==9.1.1
python-pptx==0.6.21
XlsxWriter==3.0.3
